Clues now will Destroy instead of disable / Bug fixed
there was a bug when journal was open again when you clicking on clues. it would show clueinspectionPanel
