+++ Assets/CoreSystems/Journal/JournalBackGroundIMG.prefab
@@ -3253,7 +3253,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &8533245511157011613
 RectTransform:
   m_ObjectHideFlags: 0
