+++ Assets/CoreSystems/Storage/Scripts/Clue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be5ec6a65e909e3439047204be51e2fb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
