+++ Assets/CoreSystems/Storage/Scripts/Clue/Clue.cs
@@ -7,15 +7,7 @@ public class Clue : InteractableObject
 {
     [SerializeField] private ClueData clueData;
     [SerializeField] private GameObject clueUIPanel;
-    [SerializeField] private GameObject enlargedCluePanel;
-    [SerializeField] private Animation enlargeAnimation; //Inspection Zoom, Not working/decided yet
-    [SerializeField] private GameObject clueInspectionPanel;
-
-    protected override void Start()
-    {
-        base.Start();
-    }
-
+    
     public override void Interact(InteractionOptionInstance option)
     {
         base.Interact(option);
@@ -24,34 +16,16 @@ public class Clue : InteractableObject
 
     private void CollectClue() //it will destroy the 3D item and set the sprite data to the assigned CLueUIPanel
     {
+        ClueManager.Instance.SetCurrentClue(clueData);
+        
         if (clueUIPanel != null)
         {
             clueUIPanel.GetComponent<Image>().sprite = clueData.sprite; 
             clueUIPanel.SetActive(true); 
         }
         
-        gameObject.SetActive(false);
-        
+        Destroy(gameObject);
     }
-
-    public void InspectClue() //it will assign the sprite to the deactived big panel on the ui and activate the panel
-    { //zooming
-        if (enlargedCluePanel != null)
-        {
-            enlargedCluePanel.GetComponent<Image>().sprite = clueData.sprite;
-            clueInspectionPanel.SetActive(true);
-            if (enlargeAnimation != null)
-            {
-                enlargeAnimation.Play("ClueEnlargeAnimation"); 
-            }
-        }
-    }
-
-    public void CloseInspection()
-    {
-        clueInspectionPanel.SetActive(false);
-    }
-
     public override List<InteractionOptionInstance> GetAvailabeleOptions()
     {
         return new List<InteractionOptionInstance>()
