+++ Assets/CoreSystems/Storage/Scripts/Clue/ClueManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f22be5b1446993f428bf1d0acfb5c1bd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
