+++ Assets/GameplayObjects/Note/NoteClue.asset
@@ -14,4 +14,4 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   clueName: Note
   clueDescription: MorseCode
-  sprite: {fileID: 21300000, guid: 86adaa715d8969e48975ca3343ddaedd, type: 3}
+  sprite: {fileID: 21300000, guid: 4bfa4cdfde8dae14eb6c5a34f2038240, type: 3}
