+++ Assets/GameplayObjects/UI/CommonUI.prefab
@@ -1,5 +1,55 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
+--- !u!1 &348746720385858081
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1590976827284349836}
+  - component: {fileID: 1132373715835384640}
+  m_Layer: 0
+  m_Name: ClueManager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!224 &1590976827284349836
+RectTransform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 348746720385858081}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 2955176470212897591}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_AnchorMin: {x: 0.5, y: 0.5}
+  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchoredPosition: {x: 0, y: 0}
+  m_SizeDelta: {x: 100, y: 100}
+  m_Pivot: {x: 0.5, y: 0.5}
+--- !u!114 &1132373715835384640
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 348746720385858081}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f22be5b1446993f428bf1d0acfb5c1bd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  enlargedCluePanel: {fileID: 6604375106693979688}
+  clueInspectionPanel: {fileID: 820421777205693133}
 --- !u!1 &820421777205693133
 GameObject:
   m_ObjectHideFlags: 0
@@ -12,7 +62,7 @@ GameObject:
   - component: {fileID: 4919790492596297796}
   - component: {fileID: 218118822629357166}
   m_Layer: 5
-  m_Name: ClueInspection
+  m_Name: ClueInspectionPanel
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -671,8 +721,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 6258920724349033386, guid: ab32650b7adce5f4684c2d9ea14d9dbf, type: 3}
-        m_TargetAssemblyTypeName: Clue, Assembly-CSharp
+      - m_Target: {fileID: 1132373715835384640}
+        m_TargetAssemblyTypeName: ClueManager, Assembly-CSharp
         m_MethodName: CloseInspection
         m_Mode: 1
         m_Arguments:
@@ -1252,9 +1302,9 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 0}
-        m_TargetAssemblyTypeName: Clue, Assembly-CSharp
-        m_MethodName: InspectClue
+      - m_Target: {fileID: 1132373715835384640}
+        m_TargetAssemblyTypeName: ClueManager, Assembly-CSharp
+        m_MethodName: InspectCurrentClue
         m_Mode: 1
         m_Arguments:
           m_ObjectArgument: {fileID: 0}
@@ -1806,6 +1856,9 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 7663495962766200259, guid: d87243edab6b67b42adb7705b02d171a, type: 3}
       insertIndex: -1
       addedObject: {fileID: 1342903097802543988}
+    - targetCorrespondingSourceObject: {fileID: 7663495962766200259, guid: d87243edab6b67b42adb7705b02d171a, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1590976827284349836}
     m_AddedComponents:
     - targetCorrespondingSourceObject: {fileID: 4996990604765026530, guid: d87243edab6b67b42adb7705b02d171a, type: 3}
       insertIndex: -1
